requirements.txt
einops==0.8.1
geopy==2.4.1
joblib==1.2.0
numpy==1.24.2
pandas==1.3.5
PyYAML==6.0
PyYAML==6.0.2
scipy==1.16.0
torch==2.0.0
torch_geometric==2.3.0
torch_scatter==2.1.1+pt20cu118
